+++ packages/path_provider/path_provider_ohos/CHANGELOG.md
@@ -1,3 +1,8 @@
+## path_provider-v2.1.5-ohos-1.0.0 - 2026.8
+
+**Changed**
+* Upgrade path_provider_ohos to Flutter 3.44.([#1595](https://gitcode.com/CPF-Flutter/flutter_packages/pull/1595))
+
 ## 1.0.0
 
 * new ohos version.